poller/pyproject.toml
[project]
name = "poller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.52.0",
    "cohere>=5.15.0",
    "minio>=7.2.15",
    "mypy>=1.15.0",
    "pgvector>=0.4.1",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.0",
    "types-psycopg2>=2.9.21.20250516",
]
